+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added import keyword to import templates into other templates
+
 ### Changed
 
 - Log error instead of warning when a template file is not found